Moved files to packages and Maven structure
