+++ VirtualComputer/dependency-reduced-pom.xml
@@ -5,22 +5,6 @@
   <artifactId>VirtualComputer</artifactId>
   <version>2.0-SNAPSHOT</version>
   <build>
-    <sourceDirectory>src</sourceDirectory>
-    <resources>
-      <resource>
-        <directory>src</directory>
-        <excludes>
-          <exclude>**/*.java</exclude>
-        </excludes>
-      </resource>
-      <resource>
-        <directory>.</directory>
-        <includes>
-          <include>*.dll</include>
-          <include>*.yml</include>
-        </includes>
-      </resource>
-    </resources>
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
@@ -28,9 +12,9 @@
         <configuration>
           <source>1.8</source>
           <target>1.8</target>
-            <compilerArgs>
-                <arg>-g</arg>
-            </compilerArgs>
+          <compilerArgs>
+            <arg>-g</arg>
+          </compilerArgs>
         </configuration>
       </plugin>
       <plugin>
@@ -90,12 +74,6 @@
           <groupId>net.md-5</groupId>
         </exclusion>
       </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.bukkit</groupId>
-      <artifactId>craftbukkit</artifactId>
-      <version>1.12.2-R0.1-SNAPSHOT</version>
-      <scope>provided</scope>
     </dependency>
   </dependencies>
 </project>
