--- VirtualComputer/plugin.yml
@@ -1,7 +1,0 @@
-name: VirtualComputer
-main: sznp.virtualcomputer.PluginMain
-version: 2.0
-commands:
-    computer:
-        usage: Use /computer start|stop|reset|key|mouse|input|fix
-        aliases: c
