+++ VirtualComputer/pom.xml
@@ -5,22 +5,6 @@
 	<artifactId>VirtualComputer</artifactId>
 	<version>2.0-SNAPSHOT</version>
 	<build>
-		<sourceDirectory>src</sourceDirectory>
-		<resources>
-			<resource>
-				<directory>src</directory>
-				<excludes>
-					<exclude>**/*.java</exclude>
-				</excludes>
-			</resource>
-			<resource>
-				<directory>.</directory>
-				<includes>
-					<include>*.dll</include>
-					<include>*.yml</include>
-				</includes>
-			</resource>
-		</resources>
 		<plugins>
 			<plugin>
 				<artifactId>maven-compiler-plugin</artifactId>
