+++ VirtualComputer/src/main/java/sznp/virtualcomputer/Commands.java
