+++ VirtualComputer/src/main/java/sznp/virtualcomputer/util/PXCLib.java
@@ -1,4 +1,4 @@
-package sznp.virtualcomputer;
+package sznp.virtualcomputer.util;
 
 import jnr.ffi.Pointer;
