+++ VirtualComputer/src/main/java/sznp/virtualcomputer/util/VBoxLib.java
@@ -1,4 +1,4 @@
-package sznp.virtualcomputer;
+package sznp.virtualcomputer.util;
 
 public interface VBoxLib {
 	void RTR3InitExe(int argc, String argv, int somethingzero);
